adjusting ddlManager

+++ core/src/main/java/org/polypheny/db/catalog/entity/CatalogDataPlacement.java
@@ -42,7 +42,7 @@ public class CatalogDataPlacement implements CatalogObject {
 
     private static final long serialVersionUID = 5192378654968316873L;
     public long tableId;
-    public int adapterId;
+    public long adapterId;
 
     public PlacementType placementType;
